Moved Quartz jobs to console app
