+++ BackgroundJobs/BackgroundJobs.csproj
@@ -1,0 +1,12 @@
+﻿<Project Sdk="Microsoft.NET.Sdk">
+
+  <PropertyGroup>
+    <OutputType>Exe</OutputType>
+    <TargetFramework>netcoreapp3.1</TargetFramework>
+  </PropertyGroup>
+
+  <ItemGroup>
+    <ProjectReference Include="..\MedicalAppointmentApp\MedicalAppointmentApp.csproj" />
+  </ItemGroup>
+
+</Project>
