+++ BackgroundJobs/EmailSending/SendMailJob.cs
@@ -7,23 +7,26 @@ using SendGrid.Helpers.Mail;
 using System;
 using System.Threading.Tasks;
 
-namespace MedicalAppointmentApp.Scheduler
+namespace BackgroundJobs.EmailSending
 {
     public class SendMailJob : IJob
     {
-        private readonly ILogger _logger;
         private readonly ISendGridClient _sendGridClient;
         private readonly ApplicationDbContext _context;
+        private readonly ILogger<SendMailJob> _logger;
 
-        public SendMailJob(ISendGridClient sendGridClient, ILogger<SendMailJob> logger, ApplicationDbContext context)
+        public SendMailJob(ApplicationDbContext context, ISendGridClient sendGridClient, ILogger<SendMailJob> logger)
         {
             _sendGridClient = sendGridClient;
-            _logger = logger;
             _context = context;
+            _logger = logger;
         }
 
         public async Task Execute(IJobExecutionContext context)
         {
+
+            //Console.WriteLine("HelloJob is executing.");
+
             var emailsDetails = await UpcomingAppointments.GetEmailsToSend(_context);
             var senderEmail = new EmailAddress("medicalapp255@gmail.com", "MedicalApp");
 
@@ -42,7 +45,8 @@ namespace MedicalAppointmentApp.Scheduler
                     var response = await _sendGridClient.SendEmailAsync(message);
 
                     // If email is not received, use this URL to debug: https://app.sendgrid.com/email_activity 
-                    _logger.LogInformation(response.IsSuccessStatusCode ? "Email to " + emailDetails.Email + " queued successfully!" : "Something went wrong!");
+                    _logger.LogInformation(response.IsSuccessStatusCode ? "Email to " + emailDetails.Email 
+                        + " queued successfully!" : "Something went wrong!");
                 }
             }
         }
