+++ MedicalAppointmentApp.sln
@@ -1,9 +1,11 @@
 ﻿
 Microsoft Visual Studio Solution File, Format Version 12.00
-# Visual Studio Version 16
-VisualStudioVersion = 16.0.32106.194
+# Visual Studio Version 17
+VisualStudioVersion = 17.1.32210.238
 MinimumVisualStudioVersion = 10.0.40219.1
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "MedicalAppointmentApp", "MedicalAppointmentApp\MedicalAppointmentApp.csproj", "{706DCFBC-173E-4902-BDFF-DD51A169B207}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "BackgroundJobs", "BackgroundJobs\BackgroundJobs.csproj", "{24E20143-1622-40AA-A419-B7A9E2A6E764}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -15,6 +17,10 @@ Global
 		{706DCFBC-173E-4902-BDFF-DD51A169B207}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{706DCFBC-173E-4902-BDFF-DD51A169B207}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{706DCFBC-173E-4902-BDFF-DD51A169B207}.Release|Any CPU.Build.0 = Release|Any CPU
+		{24E20143-1622-40AA-A419-B7A9E2A6E764}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{24E20143-1622-40AA-A419-B7A9E2A6E764}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{24E20143-1622-40AA-A419-B7A9E2A6E764}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{24E20143-1622-40AA-A419-B7A9E2A6E764}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
