+++ MedicalAppointmentApp/Views/Appointment/UserAppointmentList.cshtml
@@ -6,6 +6,8 @@
 
 <link href="https://maxcdn.bootstrapcdn.com/font-awesome/4.7.0/css/font-awesome.min.css" rel="stylesheet" />
 
+<h1>My appointments</h1>
+
 @if(statusMsg != null)
 {
     <partial name="_StatusMessage" for="@statusMsg" />
